requirements.txt
Django==1.3.1
PyRSS2Gen==1.0.0
-e hg+https://bitbucket.org/ronaldoussoren/altgraph#egg=altgraph
distribute==0.6.24
django-tastypie==0.9.11
feedparser==5.1
gdata==2.0.16
-e git+git://github.com/dustin/py-github.git#egg=github
ipython==0.12
mechanize==0.2.5
mimeparse==0.1.3
python-dateutil==1.5
pytz==2011n
pyzmq-static==2.1.7.1
readline==6.2.1
ssh==1.7.11
virtualenv==1.6.3
virtualenvwrapper==2.7.1
gunicorn
-e hg+https://bitbucket.org/wnielson/django-chronograph#egg=django-chronograph
tweepy
